Auth: Resource restriction based on ownership and roles

+++ services/portal/src/main/java/com/loriepocket/rest/MealController.java
@@ -1,15 +1,16 @@
 package com.loriepocket.rest;
 
 import com.loriepocket.model.Meal;
+import com.loriepocket.model.User;
 import com.loriepocket.rest.assembler.MealResourceAssembler;
 import com.loriepocket.service.MealService;
+import com.loriepocket.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PagedResourcesAssembler;
 import org.springframework.hateoas.Link;
 import org.springframework.hateoas.PagedResources;
-import org.springframework.hateoas.mvc.ControllerLinkBuilder;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -27,47 +28,61 @@ import static org.springframework.hateoas.mvc.ControllerLinkBuilder.methodOn;
 @RequestMapping( value = "/api/user/{userId}", produces = MediaType.APPLICATION_JSON_VALUE )
 public class MealController {
     @Autowired
-    public MealService mealService;
+    private MealService mealService;
+
+    @Autowired
+    private UserService userService;
 
     @RequestMapping( method = RequestMethod.GET, value= "/meal")
-    @PreAuthorize("hasAnyRole('USER', 'ADMIN')")
-    HttpEntity<PagedResources<Meal>> loadAll(@PathVariable Long userId, Pageable pageable, PagedResourcesAssembler assembler) throws Exception {
+    @PreAuthorize("#userId == principal.id or hasAnyRole('MANAGER','ADMIN')")
+    HttpEntity<PagedResources<Meal>> loadAll(@PathVariable( name = "userId" ) Long userId, Pageable pageable, PagedResourcesAssembler assembler) throws Exception {
         Page<Meal> meals = this.mealService.findAllByUserId(userId, pageable);
         Link link = linkTo(methodOn(MealController.class).loadAll(userId, pageable, assembler)).withRel("meals");
         return new ResponseEntity<>(assembler.toResource(meals, new MealResourceAssembler(userId), link), HttpStatus.OK);
     }
 
-
-    @RequestMapping(method = RequestMethod.POST, value = "/meal")
-    @PreAuthorize("hasRole('ADMIN')")
-    public Meal saveUser(@PathVariable Long userId, @RequestBody Meal payload) {
-        return this.mealService.saveOrUpdate(payload);
-    }
-
     @RequestMapping(method = RequestMethod.GET, value = "/meal/{mealId}")
-    @PreAuthorize("hasRole('ADMIN')")
+    @PreAuthorize("#userId == principal.id or hasAnyRole('MANAGER','ADMIN')")
     public Meal getMeal(@PathVariable Long userId, @PathVariable Long mealId) {
-        // Find user
         return this.mealService.findById(userId);
     }
 
+    @RequestMapping(method = RequestMethod.POST, value = "/meal")
+    @PreAuthorize("#userId == principal.id or hasAnyRole('MANAGER','ADMIN')")
+    public Meal saveUser(@PathVariable Long userId, @RequestBody Meal payload) {
+        User user = this.userService.findById(userId);
+        if(user == null)
+            throw new IllegalArgumentException("Could not find a user with id :" + userId);
+
+        payload.setUser(user);
+        return this.mealService.saveOrUpdate(payload);
+    }
+
     @RequestMapping(method = RequestMethod.PUT, value = "/meal/{mealId}")
-    @PreAuthorize("hasRole('ADMIN')")
+    @PreAuthorize("#userId == principal.id or hasAnyRole('MANAGER','ADMIN')")
     public Meal updateUser(@PathVariable Long userId, @PathVariable Long mealId, @RequestBody Meal payload) {
+        User user = this.userService.findById(userId);
+        if(user == null)
+            throw new IllegalArgumentException("Could not find a user with id :" + userId);
+
         // Find user
         Meal meal = this.mealService.findById(userId);
         // Copy over the values to be updated
+        meal.setId(mealId);
         meal.setName(payload.getName());
         meal.setCalories(payload.getCalories());
         meal.setConsumedDate(payload.getConsumedDate());
+        meal.setUser(user);
         // Update user
         return this.mealService.saveOrUpdate(meal);
     }
 
     @RequestMapping(method = RequestMethod.DELETE, value = "/meal/{mealId}")
-    @PreAuthorize("hasRole('ADMIN')")
+    @PreAuthorize("#userId == principal.id or hasAnyRole('MANAGER','ADMIN')")
     public void deleteUser(@PathVariable Long userId, @PathVariable Long mealId) {
-        this.mealService.delete(mealId);
+        Meal meal = mealService.findById(mealId);
+        // Meal exist and belongs to the user stated in path
+        if(meal != null && meal.getId() == userId)
+            this.mealService.delete(mealId);
     }
-
 }
